main/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
clearbit==0.1.7
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
pkg-resources==0.0.0
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
